+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## dev
 
+- fix: ModifiedMatcher::ifModifiedSinceHeader(), ::withIfModifiedSinceHeader(), ::ifUnmodifiedSinceHeader(),
+  ::withIfUnmodifiedSinceHeader() invalid header value addition
+- fix: ETagMatcher::ifMatchHeader(), ::withIfMatchHeader(), ::ifNoneMatchHeader(), ::withIfNoneMatchHeader() invalid
+  header value addition
+- remove: ETagMatcherResult::matches() removed (BC break)
 
 ## 0.4.0